Assigns more resposibility to the daemon handler
Shift the infinite loop and the iterating to the daemon handler's
'start' method. This opens up more options for types of handlers
in the future, such as handlers that listen to ports, etc.

Also adds a abstract base class for daemon handlers and renames
the looping daemon handler class to TimedDaemonHandler.