Fixes errors with exit paths
Fixes issues with the exit path for logging errors and above so
that child process is killed no matter where in its creation point
past being forked the error orrurs.

Also adjusts the _die() function to actually die if the process is
the child, and not the calling function seeking to have the daemon
shutdown.