Enforces daemons as singletons.
The format of the Daemon bundle only stores one PID file for the
daemon, therefore it is not reliable to run multiple processes of
the same daemon at the same time unless the daemon creates its own
children, which is beyond the current scope of this bundle.

Enforces the one daemon limit even if the user runs as root or is
a different user than the one that created the daemon.

Prevents a different user from removing the PID file when they
don't have the privileges to kill the running daemon.

Prevents a different user or root from creating a second daemon
when one is already running.

Adds more logging to make it clearer what has happened when the
script fails to create or shutdown a daemon.